- Fix numpy compatibilities

+++ deep_sort_pytorch/deep_sort/sort/detection.py
@@ -27,9 +27,9 @@ class Detection(object):
     """
 
     def __init__(self, tlwh, confidence, feature, label):
-        self.tlwh = np.asarray(tlwh, dtype=np.float)
+        self.tlwh = np.asarray(tlwh, dtype=np.float64)
         self.confidence = float(confidence)
-        self.feature = np.asarray(feature, dtype=np.float32)
+        self.feature = np.asarray(feature, dtype=np.float64)
         self.label = label
 
     def to_tlbr(self):
